requirements.txt
pygame
mutagen
pdbpp
python-vlc
